Fixes #130
Moves the uuid generation event listener from the plugin to the DB API
to decouple test expectations from the plugin.
